fix(models): SQLite autoincrement 호환을 위해 BaseModel.id를 Integer로 변경

+++ notice_chat/models/base.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 from datetime import datetime
 
-from sqlalchemy import BigInteger, DateTime, func
+from sqlalchemy import Integer, DateTime, func
 from sqlalchemy.orm import DeclarativeBase, Mapped, mapped_column
 
 
@@ -16,7 +16,7 @@ class BaseModel(Base):
     __abstract__ = True
 
     id: Mapped[int] = mapped_column(
-        BigInteger,
+        Integer,
         primary_key=True,
         autoincrement=True,
         comment="내부 PK (자동 증가 정수)",
